frontend templates

+++ src/Controller/ForumController.php
@@ -2,7 +2,6 @@
 
 namespace App\Controller;
 
-use App\Repository\UserRepository;
 use App\Entity\Forum\Comment;
 use App\Form\CommentType;
 use App\Entity\Forum\Post;
@@ -17,27 +16,31 @@ use Symfony\Component\Routing\Attribute\Route;
 class ForumController extends AbstractController
 {
     #[Route('/forum', name: 'app_forum')]
-    public function index(PostRepository $postRepository, Request $request, EntityManagerInterface $entityManager, UserRepository $userRepository): Response
+    public function index(PostRepository $postRepository, Request $request, EntityManagerInterface $entityManager): Response
     {
         $post = new Post();
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            // Get the currently logged-in user
+            $user = $this->getUser();
             
-            // 1. AUTO-SETUP: Set the hidden fields
+            if (!$user) {
+                $this->addFlash('error', 'You must be logged in to create a post.');
+                return $this->redirectToRoute('app_login');
+            }
+            
+            // Set the post data
             $post->setCreatedAt(new \DateTimeImmutable());
             $post->setUpvotes(0);
-            
-            // 2. ASSIGN AUTHOR: Grab the "Test User" (ID 1) we created
-            // Later, when you have login, this will be: $this->getUser()
-            $author = $userRepository->find(1); 
-            $post->setAuthor($author);
+            $post->setAuthor($user);
 
-            // 3. SAVE
+            // Save
             $entityManager->persist($post);
             $entityManager->flush();
 
+            $this->addFlash('success', 'Post created successfully!');
             return $this->redirectToRoute('app_forum');
         }
 
@@ -46,69 +49,93 @@ class ForumController extends AbstractController
             'posts' => $postRepository->findAll(),
         ]);
     }
+
     #[Route('/forum/{id}', name: 'app_forum_show')]
-    public function show(Post $post, Request $request, EntityManagerInterface $entityManager, UserRepository $userRepository): Response
+    public function show(Post $post, Request $request, EntityManagerInterface $entityManager): Response
     {
-        // 1. Create the Comment Form
         $comment = new Comment();
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
 
-        // 2. Handle the Submit
         if ($form->isSubmitted() && $form->isValid()) {
+            // Get the currently logged-in user
+            $user = $this->getUser();
+            
+            if (!$user) {
+                $this->addFlash('error', 'You must be logged in to comment.');
+                return $this->redirectToRoute('app_login');
+            }
+
             $comment->setCreatedAt(new \DateTimeImmutable());
             $comment->setIsSolution(false);
-            $comment->setPost($post); // Link to THIS post
-            
-            // Temporary: Use User ID 1 again (until login is ready)
-            // make sure your User repository is correct!
-            $author = $userRepository->find(1); 
-            $comment->setAuthor($author);
+            $comment->setPost($post);
+            $comment->setAuthor($user);
 
             $entityManager->persist($comment);
             $entityManager->flush();
 
-            // Refresh the page to show the new comment
+            $this->addFlash('success', 'Comment added successfully!');
             return $this->redirectToRoute('app_forum_show', ['id' => $post->getId()]);
         }
-        
 
-        // 3. Render the page
         return $this->render('forum/show.html.twig', [
             'post' => $post,
             'form' => $form->createView(),
         ]);
     }
 
-    #[Route('/forum/delete/{id}', name: 'app_forum_delete')]
+    #[Route('/forum/delete/{id}', name: 'app_forum_delete', methods: ['POST'])]
     public function delete(Post $post, EntityManagerInterface $entityManager): Response
     {
-        // 1. Remove the post
+        $user = $this->getUser();
+        
+        if (!$user) {
+            $this->addFlash('error', 'You must be logged in.');
+            return $this->redirectToRoute('app_login');
+        }
+
+        // Check if user is the author or admin
+        if ($post->getAuthor()->getId() !== $user->getId() && $user->getRole() !== 'ROLE_ADMIN') {
+            $this->addFlash('error', 'You cannot delete this post.');
+            return $this->redirectToRoute('app_forum');
+        }
+
         $entityManager->remove($post);
         $entityManager->flush();
 
-        // 2. Go back to the list
+        $this->addFlash('success', 'Post deleted successfully!');
         return $this->redirectToRoute('app_forum');
     }
+
     #[Route('/forum/edit/{id}', name: 'app_forum_edit')]
     public function edit(Post $post, Request $request, EntityManagerInterface $entityManager): Response
     {
-        // 1. Create the form using the EXISTING post
+        $user = $this->getUser();
+        
+        if (!$user) {
+            $this->addFlash('error', 'You must be logged in.');
+            return $this->redirectToRoute('app_login');
+        }
+
+        // Check if user is the author or admin
+        if ($post->getAuthor()->getId() !== $user->getId() && $user->getRole() !== 'ROLE_ADMIN') {
+            $this->addFlash('error', 'You cannot edit this post.');
+            return $this->redirectToRoute('app_forum');
+        }
+
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
 
-        // 2. Handle the save
         if ($form->isSubmitted() && $form->isValid()) {
-            // We don't need persist(), just flush() because the post is already in DB
             $entityManager->flush();
 
-            // Redirect back to the post
+            $this->addFlash('success', 'Post updated successfully!');
             return $this->redirectToRoute('app_forum_show', ['id' => $post->getId()]);
         }
 
-        // 3. Render the edit page
         return $this->render('forum/edit.html.twig', [
             'form' => $form->createView(),
+            'post' => $post,
         ]);
     }
 }
